--- a/internal/xxh32/xxh32zero_test.go
+++ b/internal/xxh32/xxh32zero_test.go
@@ -55,7 +55,7 @@ func TestZeroData(t *testing.T) {
 			t.Fatalf("got %x; want %x", got, want)
 		}
 		if got, want := xxh32.ChecksumZero(data), td.sum; got != want {
-			t.Fatalf("got %x; want %x", got, want)
+			t.Errorf("got %x; want %x", got, want)
 		}
 	}
 }
@@ -92,7 +92,7 @@ func TestZeroChecksum(t *testing.T) {
 		data := []byte(td.data)
 		h := xxh32.ChecksumZero(data)
 		if got, want := h, td.sum; got != want {
-			t.Fatalf("got %x; want %x", got, want)
+			t.Errorf("got %x; want %x", got, want)
 		}
 	}
 }
@@ -103,9 +103,46 @@ func TestZeroReset(t *testing.T) {
 		_, _ = xxh.Write([]byte(td.data))
 		h := xxh.Sum32()
 		if got, want := h, td.sum; got != want {
-			t.Fatalf("got %x; want %x", got, want)
+			t.Errorf("got %x; want %x", got, want)
 		}
 		xxh.Reset()
+	}
+}
+
+func TestNil(t *testing.T) {
+	want := xxh32.ChecksumZero([]byte(""))
+
+	var xxh xxh32.XXHZero
+	xxh.Write(nil)
+	got := xxh.Sum32()
+	if got != want {
+		t.Errorf("got %x; want %x", got, want)
+	}
+
+	got = xxh32.ChecksumZero(nil)
+	if got != want {
+		t.Errorf("got %x; want %x", got, want)
+	}
+}
+
+func TestUnaligned(t *testing.T) {
+	zeros := make([]byte, 100)
+	ha := xxh32.ChecksumZero(zeros[:len(zeros)-1])
+	hu := xxh32.ChecksumZero(zeros[1:])
+	if ha != hu {
+		t.Errorf("mismatch: %x != %x", ha, hu)
+	}
+
+	var xxh xxh32.XXHZero
+	xxh.Write(zeros[:len(zeros)-1])
+	ha = xxh.Sum32()
+
+	xxh.Reset()
+	xxh.Write(zeros[1:])
+	hu = xxh32.ChecksumZero(zeros[1:])
+
+	if ha != hu {
+		t.Errorf("mismatch: %x != %x", ha, hu)
 	}
 }
 
@@ -126,6 +163,22 @@ func Benchmark_XXH32(b *testing.B) {
 func Benchmark_XXH32_Checksum(b *testing.B) {
 	for n := 0; n < b.N; n++ {
 		xxh32.ChecksumZero(testdata1)
+	}
+}
+
+// The following two benchmark the case where 3/4 calls are not 4-byte-aligned.
+func Benchmark_XXH32Unaligned(b *testing.B) {
+	var h xxh32.XXHZero
+	for n := 0; n < b.N; n++ {
+		_, _ = h.Write(testdata1[n%4:])
+		h.Sum32()
+		h.Reset()
+	}
+}
+
+func Benchmark_XXH32_ChecksumUnaligned(b *testing.B) {
+	for n := 0; n < b.N; n++ {
+		xxh32.ChecksumZero(testdata1[n%4:])
 	}
 }