Second take on an ARM assembly version of xxhash32 (#89), now with proper handling of unaligned access. It turns out that

* ARMv7 allows unaligned access for 4-byte loads, unless in compatibility mode,
* unaligned load-multiple is never allowed, but
* the Linux kernel [catches alignment exceptions and fixes them](https://www.kernel.org/doc/html/v5.5/arm/mem_alignment.html). It's surprisingly fast at doing so.

This code is as careful as can be, so it should even run on ARMv5, which is the minimum version Go supports. It has been tested with alignment fixup turned off, in which case the previous code crashed.

Benchmark results on Raspberry Pi 4B compared to old code (with GOARM=7):

```
name                old time/op    new time/op    delta
pkg:github.com/pierrec/lz4/v4/internal/xxh32 goos:linux goarch:arm
_XXH32-4               811ns ± 1%     284ns ± 2%   -65.05%  (p=0.008 n=5+5)
_XXH32_Checksum-4      736ns ± 1%     178ns ± 2%   -75.87%  (p=0.008 n=5+5)

name                old speed      new speed      delta
pkg:github.com/pierrec/lz4/v4 goos:linux goarch:arm
UncompressPg1661-4   138MB/s ± 1%   159MB/s ± 4%   +15.87%  (p=0.008 n=5+5)
UncompressDigits-4   272MB/s ± 2%   405MB/s ± 1%   +49.14%  (p=0.008 n=5+5)
UncompressTwain-4    142MB/s ± 2%   169MB/s ± 2%   +18.48%  (p=0.008 n=5+5)
UncompressRand-4     478MB/s ± 1%  1152MB/s ± 2%  +141.16%  (p=0.008 n=5+5)
```

The contrast between aligned and unaligned speed is stark, but unaligned access is still 22% and 30% better than baseline:

```
name                        time/op
pkg:github.com/pierrec/lz4/v4/internal/xxh32 goos:linux goarch:arm
_XXH32-4                       284ns ± 2%
_XXH32_Checksum-4              178ns ± 2%
_XXH32Unaligned-4              636ns ± 2%
_XXH32_ChecksumUnaligned-4     515ns ± 2%
```

Performance on amd64 is unchanged, despite the additional function call due to code restructuring.